Changed monster health from int to double; this should allow for better handling of the Tower damage multiplier (based on the Tower and Monster types)

+++ src/Tower.java
@@ -103,7 +103,7 @@ public abstract class Tower {
 			int dRow = target.getRow() - row;
 			rotation = (-1 * Math.atan2(dCol, dRow)) + (Math.PI / 2);
 
-			target.hurt((int)(attackValue * getAttackMultiplier(target.getType())));
+			target.hurt(attackValue * getAttackMultiplier(target.getType()));
 
 			framesSinceLastAttack = 0;
 		}
